Refactor graph indexing to allow streetIndex recreation sometimes
