+++ src/main/java/org/opentripplanner/analyst/cluster/ClusterGraphBuilder.java
@@ -5,7 +5,6 @@ import com.amazonaws.services.s3.model.S3Object;
 import org.apache.commons.io.IOUtils;
 import org.opentripplanner.graph_builder.GraphBuilder;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.impl.DefaultStreetVertexIndexFactory;
 import org.opentripplanner.standalone.CommandLineParameters;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -91,7 +90,7 @@ public class ClusterGraphBuilder {
         graphBuilder.run();
         Graph graph = graphBuilder.getGraph();
         graph.routerId = graphId;
-        graph.index(new DefaultStreetVertexIndexFactory());
+        graph.index(true);
         graph.index.clusterStopsAsNeeded();
         this.currGraphId = graphId;
         this.currGraph = graph;
