+++ src/main/java/org/opentripplanner/routing/graph/Graph.java
@@ -34,8 +34,6 @@ import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.edgetype.EdgeWithCleanup;
 import org.opentripplanner.routing.edgetype.StreetEdge;
 import org.opentripplanner.routing.edgetype.TripPattern;
-import org.opentripplanner.routing.impl.DefaultStreetVertexIndexFactory;
-import org.opentripplanner.routing.services.StreetVertexIndexFactory;
 import org.opentripplanner.routing.services.StreetVertexIndexService;
 import org.opentripplanner.routing.services.notes.StreetNotesService;
 import org.opentripplanner.routing.trippattern.Deduplicator;
@@ -680,36 +678,41 @@ public class Graph implements Serializable {
             IOException {
         return load(new ObjectInputStream(is), level);
     }
-
-    /**
-     * Default load. Uses DefaultStreetVertexIndexFactory.
-     * @param in
-     * @param level
-     * @return
-     * @throws IOException
-     * @throws ClassNotFoundException
-     */
-    public static Graph load(ObjectInputStream in, LoadLevel level) throws IOException,
-            ClassNotFoundException {
-        return load(in, level, new DefaultStreetVertexIndexFactory());
-    }
     
     /** 
-     * Perform indexing on vertices, edges, and timetables, and create transient data structures.
-     * This used to be done in readObject methods upon deserialization, but stand-alone mode now
-     * allows passing graphs from graphbuilder to server in memory, without a round trip through
-     * serialization. 
-     * TODO: do we really need a factory for different street vertex indexes?
+     * Perform indexing on vertices, edges, and timetables, and create transient data structures. This used to be done
+     * in readObject methods upon deserialization, but stand-alone mode now allows passing graphs from graphbuilder to
+     * server in memory, without a round trip through serialization.
+     *
+     * After a refactor in the year 2019, the method signature was changed to no longer take in an argument of a
+     * StreetVertexIndexFactory as a single factory was always being used to create the same StreetVertexIndexService.
+     * Instead the management of instances of StreetVertexIndexServices is encapsulated within this method only. Also,
+     * the parameter of `recreateStreetIndex` was added.
+     *
+     * In a lot of cases, especially when this method is called from certain build phases, the streetIndex does not need
+     * to be recreated. However, in other use cases such as the "build graph over wire" and in-memory graph builds where
+     * a newly created graph is used immediately instead of it being saved and the program exiting, a reindex is needed
+     * in order to make sure that every non-transit edge with geographic data is available in the respective indexes.
+     * In those use cases, the streetIndex was created during the graph building phase for the purposes of splitting
+     * various edges to insert certain items like transit stops. However, the graph does need to be reindexed after all
+     * items are inserted in order to have full knowledge of all items with geography such as SimpleTransfers and
+     * StreetTransitLinks. From looking at the code paths of where a new index is requested, it seems that the indexing
+     * takes place in a manner that would not result in two different StreetSplitters being used at once as the old one
+     * is typically used only during graph building.
+     *
+     * However, due to the potential for problems to arise with two StreetVertexIndexServices and StreetSplitters being
+     * active at once, a warning is logged just in case.
      */
-    public void index (StreetVertexIndexFactory indexFactory) {
-        if (streetIndex != null) {
-            throw new UnsupportedOperationException(
-                "A streetIndex has already been defined. Can't you just reuse the existing one?"
-            );
+    public void index (boolean recreateStreetIndex) {
+        if (streetIndex == null || recreateStreetIndex) {
+            if (streetIndex != null && recreateStreetIndex) {
+                LOG.warn("Overwritting an existing streetIndex! This could lead to problems if both the old and new streetIndex are used. Make sure only the new streetIndex is used going forward.");
+            }
+            LOG.info("building street index");
+            streetIndex = new StreetVertexIndexService(this);
+            LOG.info("street index built");
         }
-        streetIndex = indexFactory.newIndex(this);
-        LOG.debug("street index built.");
-        LOG.debug("Rebuilding edge and vertex indices.");
+        LOG.info("Rebuilding edge and vertex indices");
         rebuildVertexAndEdgeIndices();
         Set<TripPattern> tableTripPatterns = Sets.newHashSet();
         for (PatternArriveVertex pav : Iterables.filter(this.getVertices(), PatternArriveVertex.class)) {
@@ -720,20 +723,19 @@ public class Graph implements Serializable {
         }
         // TODO: Move this ^ stuff into the graph index
         this.index = new GraphIndex(this);
+        LOG.info("Done rebuilding edge and vertex indices");
     }
     
     /**
-     * Loading which allows you to specify StreetVertexIndexFactory and inject other implementation.
+     * Loading which allows you to inject other implementation.
      * @param in
      * @param level
-     * @param indexFactory
      * @return
      * @throws IOException
      * @throws ClassNotFoundException
      */
     @SuppressWarnings("unchecked")
-    public static Graph load(ObjectInputStream in, LoadLevel level,
-            StreetVertexIndexFactory indexFactory) throws IOException, ClassNotFoundException {
+    public static Graph load(ObjectInputStream in, LoadLevel level) throws IOException, ClassNotFoundException {
         try {
             Graph graph = (Graph) in.readObject();
             LOG.debug("Basic graph info read.");
@@ -753,7 +755,7 @@ public class Graph implements Serializable {
             }
 
             LOG.info("Main graph read. |V|={} |E|={}", graph.countVertices(), graph.countEdges());
-            graph.index(indexFactory);
+            graph.index(false);
 
             if (level == LoadLevel.FULL) {
                 return graph;
