+++ src/main/java/org/opentripplanner/routing/impl/InputStreamGraphSource.java
@@ -8,7 +8,6 @@ import com.google.common.io.ByteStreams;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Graph.LoadLevel;
 import org.opentripplanner.routing.services.GraphSource;
-import org.opentripplanner.routing.services.StreetVertexIndexFactory;
 import org.opentripplanner.standalone.Router;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -47,9 +46,6 @@ public class InputStreamGraphSource implements GraphSource {
      * The current used input stream implementation for getting graph data source.
      */
     private Streams streams;
-
-    // TODO Why do we need a factory? There is a single one implementation.
-    private StreetVertexIndexFactory streetVertexIndexFactory = new DefaultStreetVertexIndexFactory();
 
     /**
      * @param routerId
@@ -193,8 +189,7 @@ public class InputStreamGraphSource implements GraphSource {
         try (InputStream is = streams.getGraphInputStream()) {
             LOG.info("Loading graph...");
             try {
-                newGraph = Graph.load(new ObjectInputStream(is), loadLevel,
-                        streetVertexIndexFactory);
+                newGraph = Graph.load(new ObjectInputStream(is), loadLevel);
             } catch (Exception ex) {
                 LOG.error("Exception while loading graph '{}'.", routerId, ex);
                 return null;
