--- src/main/java/org/opentripplanner/routing/services/StreetVertexIndexFactory.java
@@ -1,19 +1,0 @@
-package org.opentripplanner.routing.services;
-
-import org.opentripplanner.routing.graph.Graph;
-
-
-/**
- * Factory for StreetVertexIndexServices.
- * 
- * @author avi
- */
-public interface StreetVertexIndexFactory {
-    
-    /**
-     * Returns a new StreetVertexIndexService for this graph.
-     * @param g
-     * @return
-     */
-    public StreetVertexIndexService newIndex(Graph g);
-}
