+++ src/test/java/org/opentripplanner/analyst/scenario/AddTripPatternTest.java
@@ -15,7 +15,6 @@ import org.opentripplanner.routing.core.RoutingRequest;
 import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
-import org.opentripplanner.routing.impl.DefaultStreetVertexIndexFactory;
 import org.opentripplanner.routing.spt.ShortestPathTree;
 
 import java.time.DayOfWeek;
