+++ src/test/java/org/opentripplanner/graph_builder/module/FakeGraph.java
@@ -13,7 +13,6 @@ import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.openstreetmap.impl.AnyFileBasedOpenStreetMapProviderImpl;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.impl.DefaultStreetVertexIndexFactory;
 import org.opentripplanner.routing.vertextype.TransitStop;
 
 import java.io.File;
@@ -559,7 +558,7 @@ public class FakeGraph {
      */
     public static void indexGraphAndLinkStations (Graph g) {
         // creating a new DefaultStreetVertexIndexFactory results in setting the streetIndex on the graph.
-        g.index(new DefaultStreetVertexIndexFactory());
+        g.index(false);
         StreetSplitter linker = (StreetSplitter) g.streetIndex.getStreetSplitter();
         linker.linkAllStationsToGraph();
     }
