+++ src/test/java/org/opentripplanner/routing/core/RoutingContextDestroyTest.java
@@ -14,7 +14,6 @@ import org.opentripplanner.routing.edgetype.TemporaryEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
-import org.opentripplanner.routing.impl.DefaultStreetVertexIndexFactory;
 import org.opentripplanner.routing.vertextype.IntersectionVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.TemporaryVertex;
@@ -55,7 +54,7 @@ public class RoutingContextDestroyTest {
         createStreetEdge(a, b, "a -> b");
         createStreetEdge(b, a, "b -> a");
         createStreetEdge(a, c, "a -> c");
-        g.index(new DefaultStreetVertexIndexFactory());
+        g.index(false);
     }
 
     @Test public void temporaryChangesRemovedOnContextDestroy() {
